requirements.txt
Flask==3.0.2
gunicorn==21.2.0
google-cloud==0.34.0
google-cloud-storage==2.19.0
google-cloud-datastore==2.20.2